feat: making auUse synchronous
Now that https://github.com/angular/angular/issues/50320 is fixed,
it should be possible to change any angular signal at any time,
and there should be no need anymore to delay the call of directives.

It is important (especially for transitions) to only call the directive
after its content has been properly populated, that's why this PR uses
afterNextRender to schedule the call to the directive.