refactor(react): avoid calling useDirective when element is not in the DOM
Let's always use directly the {...useDirective(myDirective)} syntax in JSX
rather than storing the result of useDirective in a variable and using it
later. As it is not allowed to call a hook a variable number of times in
a component, some react components had to be split in smaller components.